Abstract PSF loading functionality
